DOCUMENT_CLASSIFICATION/requirements.txt
# Core Framework & Web
Flask==2.3.2
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Werkzeug==2.3.6

# Database & ORM
SQLAlchemy==2.0.19
Flask-SQLAlchemy==3.0.5
alembic==1.11.1

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.0.1
cryptography==41.0.3

# Machine Learning & NLP
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
spacy==3.7.2

# Text & Document Processing
PyPDF2==3.0.1
pdfplumber==0.9.0
python-docx==0.8.11
pytesseract==0.3.10
easyocr==1.7.0

# Image Processing
Pillow==10.0.0
opencv-python==4.8.0.74

# Data Validation & Processing
marshmallow==3.20.1
jsonschema==4.19.0

# API & Serialization
requests==2.31.0
python-dotenv==1.0.0

# File Monitoring
watchdog==3.0.0

# Logging & Monitoring
python-logging-loki==0.3.1
sentry-sdk==2.5.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-flask==1.2.0
factory-boy==3.3.0

# Code Quality
black==23.7.0
flake8==6.0.0
pylint==2.17.5
isort==5.12.0

# Development
python-dotenv==1.0.0

# Server & Deployment
gunicorn==21.2.0
python-multipart==0.0.6

# Optional: For advanced features
transformers==4.31.0  # BERT models (optional, for future upgrade)
torch==2.0.1  # PyTorch for deep learning (optional)
